extract request history into its own component
