+++ bbb-client/next.config.js
@@ -1,9 +1,12 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {}
 
+// eslint-disable-next-line no-undef
 module.exports = nextConfig
 
+// eslint-disable-next-line no-undef
 module.exports = {
+    reactStrictMode: false,
     images: {
       domains: ['images.costco-static.com'],
     },
