+++ bbb-client/src/components/user/account.tsx
@@ -2,8 +2,10 @@
 
 import React, { useState, useEffect } from "react";
 import StandardButton from './button';
-import RequestCard from "./requestCard";
+// import RequestCard from "./requestCard";
+// import RequestDisplay from "./requestDisplay";
 import { useRouter } from 'next/navigation';
+import RequestDisplay from "./requestDisplay";
 
 export interface Transaction {
     _id: string;
@@ -34,53 +36,76 @@ export interface Account {
 }
 
 function Account() {
-    const [transactions, setTransactions] = useState<Transaction[]>([]);
     const [user, setUser] = useState<Account>();
+    const [doneLoading, setDoneLoading] = useState(false);
     const router = useRouter()
 
     useEffect(() => {
+        const getUserAcctInfo = async () => {
+            try {
+              const response = await fetch("http://127.0.0.1:5000/get_acct_info", {
+                credentials: "include",
+                method: "GET",
+                headers: {
+                    "Content-Type": "application/json"
+                }
+              })
+              const user_acct = await response.json();
+    
+              if (user_acct == null) {
+                router.push('/user/login')
+              }
+              console.log(user_acct)
+              setUser(user_acct)
+              setDoneLoading(true)
+            } catch (error) {
+              console.error(error);
+            }
+        }
+
         getUserAcctInfo();
         // getUserTransactions();
     }, []);
 
-    const getUserAcctInfo = async () => {
-        try {
-          const response = await fetch("http://127.0.0.1:5000/get_acct_info", {
-            credentials: "include",
-            method: "GET",
-            headers: {
-                "Content-Type": "application/json"
-            }
-          })
-          const user_acct = await response.json();
-          
-          if (user_acct == null) {
-            router.push('/user/login')
-          }
-          console.log(user_acct)
-          setUser(user_acct)
-        } catch (error) {
-          console.error(error);
-        }
-    }
+    // const getUserAcctInfo = async () => {
+    //     try {
+    //       const response = await fetch("http://127.0.0.1:5000/get_acct_info", {
+    //         credentials: "include",
+    //         method: "GET",
+    //         headers: {
+    //             "Content-Type": "application/json"
+    //         }
+    //       })
+    //       const user_acct = await response.json();
+
+    //       if (user_acct == null) {
+    //         console.log("NOOO")
+    //         router.push('/user/login')
+    //       }
+    //       console.log(user_acct)
+    //       setUser(user_acct)
+    //     } catch (error) {
+    //       console.error(error);
+    //     }
+    // }
 
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const getUserTransactions= async () => {
-        try {
-          const response = await fetch("http://127.0.0.1:5000/get_transactions", {
-            credentials: "include",
-            method: "GET",
-            headers: {
-                "Content-Type": "application/json"
-            }
-          })
-          const requests = await response.json();
-          console.log(requests)
-          setTransactions(requests)
-        } catch (error) {
-          console.error(error);
-        }
-    }
+    // const getUserTransactions= async () => {
+    //     try {
+    //       const response = await fetch("http://127.0.0.1:5000/get_transactions", {
+    //         credentials: "include",
+    //         method: "GET",
+    //         headers: {
+    //             "Content-Type": "application/json"
+    //         }
+    //       })
+    //       const requests = await response.json();
+    //       console.log(requests)
+    //       setTransactions(requests)
+    //     } catch (error) {
+    //       console.error(error);
+    //     }
+    // }
 
     const navigateUpdateUser = async () => {
         try {
@@ -96,12 +121,19 @@ function Account() {
             <h1> Welcome, {user?.firstname} </h1>
             <p> Account: {user?.uid} </p>
             <StandardButton onClick={navigateUpdateUser} label="Edit account information" />
-            <h1> History </h1>
+
+            {doneLoading ?
+                <RequestDisplay/>
+                :
+                <div/>
+            }
+            
+            {/* <h1> History </h1>
             <ul className="product-list">
                 {transactions.map((request: Transaction) => (
                 <RequestCard request={request} key={request._id} />
                 ))}
-            </ul>
+            </ul> */}
         </div>
     )
